+++ app/helpers/folio/open_graph_helper.rb
@@ -3,7 +3,7 @@
 module Folio::OpenGraphHelper
   def og_image
     return @og_image if @og_image.present?
-    path = defined?(og_image_fallback) ? og_image_fallback : current_site.og_image_fallback
+    path = defined?(og_image_fallback) ? og_image_fallback : current_site.og_image_with_fallback
     image_url(path, host: request.base_url)
   rescue Sprockets::Rails::Helper::AssetNotFound
     nil