+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2021-01-10
+### Added
+- added `Folio::FilePlacement::OgImage` to default file plcaement types, add `Folio::HasAttachments` and update fallback og:image
+
 ## 2021-01-06
 ### Changed
 - reinvite `Folio::User` when signing in using an e-mail of an user with a pending invitation